requirements.txt
cycler==0.10.0
Django==2.1.5
django-ratelimit==2.0.0
kiwisolver==1.1.0
matplotlib==2.2.3
numpy==1.18.2
Pillow==5.2.0
pybase64==1.0.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-memcached==1.59
pytz==2019.3
six==1.14.0
